Add ctrls_io.f to the list of files read in for controls namelist
Previously MesaScript was unable to read in commands to read from additional
inlists (like those commands found in the default inlist in star/work/inlist).
These were in a separate file that is now read in addition to star_controls.inc.

Additionally fixes a problem where the extra inlist controls for pgstar were
not able to be read in. The underlying problem (indexing above the size of
the array in @to_read['pgstar']) is not really solved, but instead a workaround
(turning newly formed nil entries to be converted to empty strings) is inserted.