Fix issues with multi-d arrays and default acquisition.
Previously, multi-d arrays (like summary text blocks in pgstar) were not dealt
with in any sensible way. Now they are functioning and are aware of how many
indices are required. They can be accessed by using an array of indices as the
first argument to the relevant command OR by specifying all indices as the first
N arguments of the function. The N+1 argument can be used for setting the
relevant value. Using an array of indices also works in the hash way of setting
array values, so

>>>> text_summary1_name[1,1] = 'my_var'

will work just as well as

>>>> text_summary1_name(1, 1, 'my_var')

Also fix an issue where defaults were not stored (!). Now default values are
properly stored and remembered.